+++ R/cik_search.R
@@ -19,10 +19,7 @@
 cik_search <- function(company) {
   href <- paste0("https://www.sec.gov/cgi-bin/cik_lookup",
                 "?company=", URLencode(company, reserved = TRUE))
-  res <- httr::GET(href)
-  if (res$status != "200") {
-    stop("Unable to reach the SEC cik lookup endpoint (https://www.sec.gov/cgi-bin/cik_lookup)")
-  }
+  res <- edgar_GET(href)
   doc <- xml2::read_html(res, base_url = href, options = "HUGE")
 
   entries_xpath <- "//pre/a[starts-with(@href,'browse-edgar')]"