+++ R/utils.R
@@ -64,10 +64,7 @@ is_url <- function(x) {
 get_doc <- function(x, clean = F) {
   if (typeof(x) == "character") {
     if (is_url(x)) {
-      res <- httr::GET(x)
-      if (res$status != "200") {
-        stop(paste0("Unable to reach the SEC endpoint (", x, ")"))
-      }
+      res <- edgar_GET(x)
       content <- httr::content(res, encoding = "UTF-8", as = "text")
       if (clean) {
         content <- clean_html(content)
@@ -107,7 +104,7 @@ get_doc <- function(x, clean = F) {
 
 charToDoc <- function(x) {
   if (is_url(x)) {
-    res <- httr::GET(x)
+    res <- edgar_GET(x)
     if (res$status != "200") {
       stop(paste0("Unable to reach the SEC endpoint (", x, ")"))
     }
@@ -119,7 +116,7 @@ charToDoc <- function(x) {
 
 charToText <- function(x) {
   if (is_url(x)) {
-    res <- httr::GET(x)
+    res <- edgar_GET(x)
     if (res$status != "200") {
       stop(paste0("Unable to reach the SEC endpoint (", x, ")"))
     }