+++ tests/testthat/test_browse_edgar.R
@@ -2,8 +2,8 @@ context("running browse_edgar")
 
 with_mock_API({
   test_that("running ", {
-    expect_error(browse_edgar("EAR"),
-                 "Could not find company: EAR")
+    expect_error(browse_edgar("BAD"),
+                 "Could not find company: BAD")
     expect_s3_class(browse_edgar("EA"), "xml_node")
   })
 })