+++ tests/testthat/test_company_information.R
@@ -2,7 +2,7 @@ context("running company_information")
 
 with_mock_API({
 test_that("running ", {
-            expect_error(company_information("EAR"))
+            expect_error(company_information("BAD"))
             res <- company_information("EA")
             expect_is(res, "data.frame")